refactor: renamed layout component files and folders
