--- layout/Main/components/AppMain/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AppMain';
