--- layout/Main/components/FooterNav/index.ts
@@ -1,1 +1,0 @@
-export { default } from './FooterNav';
