--- layout/Main/components/HeaderNav/index.ts
@@ -1,1 +1,0 @@
-export { default } from './HeaderNav';
