--- layout/Main/components/Navigation/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Navigation';
