+++ layouts/Main/components/NavLinks/index.ts
@@ -1,0 +1,1 @@
+export { LinksGroup } from './NavLinks';
